Trait count adjustment
Added a new adjustment factor to the rarity score calculation that takes into account the number of trait values an NFT has. This adjustment ensures that NFTs with different numbers of trait values can be compared fairly, as an NFT with more trait values is likely to have a higher rarity score simply because it has more opportunities for rare trait values to occur.

The adjustment factor is calculated by dividing 1 by the number of trait values an NFT has, and is stored in the traitCountFactor variable. The rarity score for each NFT is then multiplied by this adjustment factor, resulting in an adjusted rarity score that reflects the NFT's rarity relative to its number of trait values.